pyproject.toml
[tool.poetry]
name = "enpkg_full"
version = "0.0.1"
description = "ENPKG FULL"
authors = ["Pierre-Marie Allard <pierre-marie.allard@unifr.ch>"]
license = "LICENSE"
packages = [{include = "*.py", from = "src/"}]
readme = "README.md"
homepage = "https://github.com/enpkg/enpkg_full"
repository = "https://github.com/enpkg/enpkg_full"
documentation = "https://github.com/enpkg/enpkg_full"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
ipykernel = "^6.28.0"
ipython = "^8.10"
matchms = "^0.20.0"
matplotlib = "^3.8.2"
networkx = "^3.2.1"
numpy = "^1.24"
pandas = "^2.0.3"
plotly = "^5.18.0"
pyyaml = "^6.0.1"
rdflib = "^7.0.0"
rdkit = "^2023.9.4"
requests = "^2.31.0"
scikit-bio = "^0.5.9"
scikit-learn = "^1.3.2"
scipy = "^1.10"
seaborn = "^0.13.1"
spec2vec = "^0.8.0"
tqdm = "^4.66.1"
bokeh = "^2.4.3"
chembl-webresource-client = "^0.10.8"
compress-json = "^1.0.10"
datatable = "^1.0.0"
downloaders = "^1.0.20"
gitpython = "^3.1.41"
memo-ms = "^0.1.4"
opentree = "^1.0.1"
userinput = "^1.0.22"
lxml = "^5.2.2"
taxonomical_utils = "^0.0.5"


[tool.poetry.dev-dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.2"
pytest-mock = "*"
requests_mock = "*"
coverage = "^7.4.0"
pytest-xdist = "^3.5.0"
pytest-cov = "^4.1.0"


[tool.poetry.scripts]
start = "app:main"

[tool.poetry.group.dev.dependencies]
tox = "^4.12.0"
black = "^23.12.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:The module rdkit.Chem.MolStandardize:DeprecationWarning"
]
testpaths = "tests"

[tool.isort]
profile = "black"